Updated MSVC projects for new classes from Adam. MrsLog now supports logging to cout (if LOG_TO_COUT is defined somewhere, e.g. common.h) small assorted fixes
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@435 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ marsyas/MarControl.cpp
@@ -184,7 +184,7 @@ MarControl::linkTo(MarControlPtr ctrl)
 void
 MarControl::clearLinks()
 {
-	for (int i=0; i<linkedTo_.size(); i++)
+	for (size_t i=0; i<linkedTo_.size(); i++)
 	{
 		linkedTo_[i]->removeLink(this);
 	}
@@ -195,7 +195,7 @@ void
 MarControl::removeLink(MarControlPtr link)
 {
 	std::vector<MarControlPtr> temp;
-	for (int i=0; i<linkedTo_.size(); i++)
+	for (size_t i=0; i<linkedTo_.size(); i++)
 	{
 		if (linkedTo_[i]() != link()) // same pointer?
 		{
